Make duty_ns an int

+++ hygencms/pwm.py
@@ -268,7 +268,7 @@ def set_frequency(pin_name, freq):
                       bytes("{:d}".format(period_ns), encoding='utf-8'))
 
         # Calculate updated duty cycle
-        duty_ns = (pin.duty / 100.) * period_ns
+        duty_ns = int((pin.duty / 100.) * period_ns)
 
         os.lseek(pin.duty_fd, 0, os.SEEK_SET)
         n2 = os.write(pin.duty_fd,
